API reworked
Now scrapper and api work separately. First run server, then you can run scrapper or make post requests to api. It saves all records on server side, not in memory.

Now it takes a json and returns a json, not a dictionary as previously.

Unnecessary code like html code or page names deleted.

PUT method is to be reworked.