delete_document() reworked, API checking if doc exists reworked.
Except simple removing a post from db, function now checks whether author has any more posts, if not it removes author as well.

Instead of calling find_document() to check whether document exists or not, now it makes request and cathes TypeError exception, if TypeError sends 404.

Server doesn't assign a unique id to post anymore. Now unique id is to be given in JSON to be post.

Docstring is to be fixed.